requirements.txt
numpy
scipy
matplotlib
tqdm
pyyaml
celluloid
pytest